requirements.txt
black==22.1.0
Pillow==9.0.1
pre-commit==2.17.0
pytest==7.0.1
pytest-cov==3.0.0
